requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
python-dotenv==1.0.1
google-generativeai==0.3.2
streamlit==1.30.0
langdetect
googletrans==4.0.0-rc1
deep-translator==1.4.2